fix/EME2-6 - seperated calculator logic from routes into own logic folder
